fix: javax GL10/EGLConfig renderer signature, LunarVRActivity type, and all Float/Double math mismatches (Quat, Mat4, HeadTracker, HandPose, VirtualHand, Environment, UI, SoundBank)

+++ app/src/main/java/com/lunarvr/audio/SoundBank.kt
@@ -70,7 +70,7 @@ class SoundBank(private val context: Context) {
     // ------------------------------------------------------------------
     private fun env(t: Float, dur: Float, attack: Float = 0.004f): Float {
         val a = min(t / attack, 1f)
-        val d = exp(-(t / dur) * 5f)
+        val d = exp((-t / dur).toDouble() * 5.0).toFloat()
         return a * d
     }
